requirements.txt
asgiref==3.5.2
Django==4.1
django-autoslug==1.9.8
django-birthday==0.1.4
django-cities-light==3.9.2
django-countries==7.3.2
progressbar2==4.0.0
python-utils==3.3.3
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
Unidecode==1.3.4
